Move MCP response suggestions from _core to _mcp response formatters
- Relocate suggestions.py to _mcp/response_formatters/.
- Update imports in tool handlers (search_tools, call_graph_tools).
- No behavior change.

make format-check, make lint, make type-check pass.
Targeted MCP tests: 35 passed.
